python-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
psycopg2-binary>=2.9.9
anthropic>=0.40.0
openai>=1.54.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pytest>=8.3.0
httpx>=0.27.0
pandas>=2.2.0
